Convert all "comment" components to typescript
